+++ Source/BasicTypes/TShortString.cpp
@@ -37,7 +37,7 @@ TShortString::TShortString(const char* pChar, unsigned short pCharLen)
 	{
 		if (pCharLen==0)
 		{
-			pCharLen = strlen(pChar);
+			pCharLen = (unsigned short)strlen(pChar);
 		}
 		CopyFrom(pChar, pCharLen);    
 	}
